Add support for inverted indexes
To use full-text search the ArangoDB docs point you towards using
an inverted index: https://docs.arangodb.com/3.11/index-and-search/arangosearch/full-text-token-search/

ArangoMigo doesn't support that at the moment so this adds that
as a migration type.

The ArangoDB Go driver had to be updated - inverted indexes look to
be a newer addition:
> Please note that the fulltext index type is deprecated from version
> 3.10 onwards and is superseded by ArangoSearch.
(https://docs.arangodb.com/3.11/index-and-search/indexing/which-index-to-use-when/)

This doesn't support all features of inverted index, just the
minimum necessary.